refactor(frontend): move `DefKind` and below to `def_id.rs`
